Fix warning about signed / unsigned comparison.

+++ DLLDependencyViewer/DLLDependencyViewer/src/nogui/pe/coff_full.cpp
@@ -28,12 +28,12 @@ bool pe_parse_coff_full_32_64(void const* const& fd, int const& file_size, coff_
 	WARN_M_R(file_size >= static_cast<int>(dosheader.m_pe_offset + sizeof(coff_full_32_64)), L"File too small to contain coff_full_32_64.", false);
 	hd = reinterpret_cast<coff_full_32_64 const*>(coff_hdr);
 	coff_full_32_64 const& header = *hd;
-	WARN_M_R(file_size >= dosheader.m_pe_offset + (is_32 ? sizeof(coff_full_32) : sizeof(coff_full_64) + dir_cnt * sizeof(data_directory)), L"File too small to contain all directories.", false);
+	WARN_M_R(file_size >= static_cast<int>(dosheader.m_pe_offset + (is_32 ? sizeof(coff_full_32) : sizeof(coff_full_64) + dir_cnt * sizeof(data_directory))), L"File too small to contain all directories.", false);
 	data_directory const* const directories = reinterpret_cast<data_directory const*>(file_data + dosheader.m_pe_offset + (is_32 ? sizeof(coff_full_32) : sizeof(coff_full_64)));
 	WARN_M(dir_cnt < static_cast<std::uint32_t>(e_directory_table::architecture) || (directories[static_cast<int>(e_directory_table::architecture)].m_va == 0 && directories[static_cast<int>(e_directory_table::architecture)].m_size == 0), L"Architecture is reserved, must be 0.");
 	WARN_M(dir_cnt < static_cast<std::uint32_t>(e_directory_table::global_ptr) || directories[static_cast<int>(e_directory_table::global_ptr)].m_size == 0, L"The size member of Global Ptr structure must be set to zero.");
 	WARN_M(dir_cnt < static_cast<std::uint32_t>(e_directory_table::reserved) || (directories[static_cast<int>(e_directory_table::reserved)].m_va == 0 && directories[static_cast<int>(e_directory_table::reserved)].m_size == 0), L"Reserved, must be zero.");
-	WARN_M_R(file_size >= dosheader.m_pe_offset + (is_32 ? sizeof(coff_full_32) : sizeof(coff_full_64)) + dir_cnt * sizeof(data_directory) + coff_hdr->m_section_count * sizeof(section_header), L"File too small to contain all section headers.", false);
+	WARN_M_R(file_size >= static_cast<int>(dosheader.m_pe_offset + (is_32 ? sizeof(coff_full_32) : sizeof(coff_full_64)) + dir_cnt * sizeof(data_directory) + coff_hdr->m_section_count * sizeof(section_header)), L"File too small to contain all section headers.", false);
 	section_header const* const sections = reinterpret_cast<section_header const*>(file_data + dosheader.m_pe_offset + (is_32 ? sizeof(coff_full_32) : sizeof(coff_full_64)) + dir_cnt * sizeof(data_directory));
 	for(std::uint16_t i = 1; i < coff_hdr->m_section_count; ++i)
 	{
